homework/hw4/241878_hw4_2025_1/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
_openmp_mutex=4.5=2_gnu
affine=2.4.0=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.3.0=pyh71513ae_0
beautifulsoup4=4.13.4=pypi_0
blosc=1.21.6=hfd34d9b_1
branca=0.8.1=pyhd8ed1ab_0
brotli=1.1.0=h2466b09_2
brotli-bin=1.1.0=h2466b09_2
brotli-python=1.1.0=py310h9e98ed7_2
bzip2=1.0.8=h2466b09_7
ca-certificates=2025.4.26=h4c7d964_0
certifi=2025.4.26=pyhd8ed1ab_0
cffi=1.17.1=py310ha8f682b_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
click=8.2.1=pyh7428d3b_0
click-plugins=1.1.1=pyhd8ed1ab_1
cligj=0.7.2=pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=py310hc19bc0b_0
cpython=3.10.17=py310hd8ed1ab_0
cycler=0.12.1=pyhd8ed1ab_1
debugpy=1.8.14=py310h9e98ed7_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
fiona=1.10.1=py310hf80745b_3
folium=0.19.6=pyhd8ed1ab_0
fonttools=4.58.1=py310h38315fa_0
freetype=2.13.3=h57928b3_1
freexl=2.0.0=hf297d47_2
gdal=3.10.3=py310hdc3ad04_10
geopandas=1.0.1=pyhd8ed1ab_3
geopandas-base=1.0.1=pyha770c72_3
geos=3.13.1=h9ea8674_0
geotiff=1.7.4=h86c3423_2
h2=4.2.0=pyhd8ed1ab_0
h5py=3.13.0=pypi_0
hdf4=4.2.15=h5557f11_7
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.7.0=pyhe01879c_1
intel-openmp=2024.2.1=h57928b3_1083
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.36.0=pyh9ab4c32_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.5.1=pyhd8ed1ab_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh5737063_0
kiwisolver=1.4.7=py310hc19bc0b_0
krb5=1.21.3=hdf4eb48_0
lcms2=2.17=hbcf6048_0
lerc=4.0.0=h6470a55_1
libarchive=3.7.7=h5343c79_4
libblas=3.9.0=31_h641d27c_mkl
libbrotlicommon=1.1.0=h2466b09_2
libbrotlidec=1.1.0=h2466b09_2
libbrotlienc=1.1.0=h2466b09_2
libcblas=3.9.0=31_h5e41251_mkl
libcurl=8.14.0=h88aaa65_0
libdeflate=1.24=h76ddb4d_0
libexpat=2.7.0=he0c23c2_0
libffi=3.4.6=h537db12_1
libfreetype=2.13.3=h57928b3_1
libfreetype6=2.13.3=h0b5ce68_1
libgcc=15.1.0=h1383e82_2
libgdal-core=3.10.3=heb20a2f_10
libgomp=15.1.0=h1383e82_2
libhwloc=2.11.2=default_ha69328c_1001
libiconv=1.18=h135ad9c_1
libjpeg-turbo=3.1.0=h2466b09_0
libkml=1.3.0=h538826c_1021
liblapack=3.9.0=31_h1aa476e_mkl
liblzma=5.8.1=h2466b09_1
libpng=1.6.47=h7a4582a_0
librttopo=1.1.0=hbfc9ebc_18
libsodium=1.0.20=hc70643c_0
libspatialite=5.1.0=h378fb81_14
libsqlite=3.50.0=h67fdade_0
libssh2=1.11.1=h9aa295b_0
libtiff=4.7.0=h05922d8_5
libwebp-base=1.5.0=h3b0e114_0
libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_9
libxcb=1.17.0=h0e4246c_0
libxml2=2.13.8=h442d1da_0
libzlib=1.3.1=h2466b09_2
lz4-c=1.10.0=h2466b09_1
lzo=2.10=hcfcfb64_1001
mapclassify=2.8.1=pyhd8ed1ab_1
markupsafe=3.0.2=py310h38315fa_1
matplotlib-base=3.10.3=py310h37e0a56_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
minizip=4.0.10=h9fa1bad_0
mkl=2024.2.2=h66d3029_15
munkres=1.1.4=pyh9f0ad1d_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pyh267e887_2
numpy=2.2.6=py310h4987827_0
openjpeg=2.5.3=h4d64b90_0
openssl=3.5.0=ha4e3fda_1
osmnx=2.0.3=ha770c72_0
osmnx-base=2.0.3=pyhd8ed1ab_0
packaging=25.0=pyh29332c3_1
pandas=2.2.3=py310hb4db72f_3
parso=0.8.4=pyhd8ed1ab_1
pcre2=10.45=h99c9b8b_0
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.2.1=py310h9595edc_0
pip=25.1.1=pyh8b19718_0
platformdirs=4.3.8=pyhe01879c_0
proj=9.6.0=h4f671f6_1
prompt-toolkit=3.0.51=pyha770c72_0
psutil=7.0.0=py310ha8f682b_0
pthread-stubs=0.4=h0e40799_1002
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.22=pyh29332c3_1
pygments=2.19.1=pyhd8ed1ab_0
pyhdf=0.11.6=py310hd61a6a9_0
pyogrio=0.11.0=py310hf80745b_0
pyparsing=3.2.3=pyhd8ed1ab_1
pyproj=3.7.1=py310h61ea1ad_1
pysocks=1.7.1=pyh09c184e_7
python=3.10.17=h8c5b53a_0_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.10=7_cp310
pytz=2025.2=pyhd8ed1ab_0
pywin32=307=py310h9e98ed7_3
pyzmq=26.4.0=py310h656833d_0
qhull=2020.2=hc790b64_5
rasterio=1.4.3=py310h5977aad_1
rasterstats=0.20.0=pyhd8ed1ab_1
requests=2.32.3=pyhd8ed1ab_1
rio-vrt=0.3.1=pyhd8ed1ab_1
rioxarray=0.19.0=pyhd8ed1ab_0
scikit-learn=1.6.1=py310hf2a6c47_0
scipy=1.15.2=py310h15c175c_0
seaborn=0.13.2=pypi_0
setuptools=80.8.0=pyhff2d567_0
shapely=2.1.1=py310hb052cae_0
simplejson=3.20.1=py310ha8f682b_0
six=1.17.0=pyhd8ed1ab_0
snappy=1.2.1=h500f7fa_1
snuggs=1.4.7=pyhd8ed1ab_2
soupsieve=2.7=pypi_0
sqlite=3.50.0=h2466b09_0
stack_data=0.6.3=pyhd8ed1ab_1
tbb=2021.13.0=h62715c5_1
threadpoolctl=3.6.0=pyhecae5ae_0
tk=8.6.13=h2c6b04d_2
tornado=6.5.1=py310ha8f682b_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
ucrt=10.0.22621.0=h57928b3_1
unicodedata2=16.0.0=py310ha8f682b_0
uriparser=0.9.8=h5a68840_0
urllib3=2.4.0=pyhd8ed1ab_0
vc=14.3=h2b53caa_26
vc14_runtime=14.42.34438=hfd919c2_26
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
win_inet_pton=1.1.0=pyh7428d3b_8
xarray=2025.4.0=pyhd8ed1ab_0
xerces-c=3.2.5=he0c23c2_2
xorg-libxau=1.0.12=h0e40799_0
xorg-libxdmcp=1.1.5=h0e40799_0
xyzservices=2025.4.0=pyhd8ed1ab_0
zeromq=4.3.5=ha9f60a1_7
zipp=3.22.0=pyhd8ed1ab_0
zlib=1.3.1=h2466b09_2
zstandard=0.23.0=py310ha8f682b_2
zstd=1.5.7=hbeecb71_2
